refactor: move SelectedContextType enum to enums folder
