+++ frontend/dashboard/context/HeaderContext/HeaderContext.tsx
@@ -3,12 +3,6 @@ import { type Organization } from 'app-shared/types/Organization';
 import { type User } from 'app-shared/types/Repository';
 import { APP_DASHBOARD_BASENAME, ORG_LIBRARY_BASENAME } from 'app-shared/constants';
 
-export enum SelectedContextType {
-  All = 'all',
-  Self = 'self',
-  None = 'none',
-}
-
 export enum Subroute {
   AppDashboard = APP_DASHBOARD_BASENAME,
   OrgLibrary = ORG_LIBRARY_BASENAME,
