+++ frontend/dashboard/enums/SelectedContextType.ts
@@ -1,0 +1,5 @@
+export enum SelectedContextType {
+  All = 'all',
+  Self = 'self',
+  None = 'none',
+}
