+++ frontend/dashboard/hooks/guards/useContextRedirectionGuard.ts
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useSelectedContext } from '../useSelectedContext';
 import type { NavigateFunction } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
-import { SelectedContextType } from 'dashboard/context/HeaderContext';
+import { SelectedContextType } from '../../enums/SelectedContextType';
 import { typedSessionStorage } from '@studio/pure-functions';
 import { userHasAccessToSelectedContext } from 'dashboard/utils/userUtils';
 import { useSubroute } from '../useSubRoute';
