+++ frontend/dashboard/hooks/usePageHeaderTitle/usePageHeaderTitle.test.tsx
@@ -3,7 +3,7 @@ import { usePageHeaderTitle } from './usePageHeaderTitle';
 import { HeaderContext, type HeaderContextType } from 'dashboard/context/HeaderContext';
 import { useSelectedContext } from 'dashboard/hooks/useSelectedContext';
 import { headerContextValueMock } from 'dashboard/testing/headerContextMock';
-import { SelectedContextType } from 'dashboard/context/HeaderContext';
+import { SelectedContextType } from '../../enums/SelectedContextType';
 import { mockOrg1 } from 'dashboard/testing/organizationMock';
 import { renderHookWithProviders } from 'dashboard/testing/mocks';
