+++ frontend/dashboard/hooks/useSelectedContext/useSelectedContext.ts
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import { SelectedContextType } from 'dashboard/context/HeaderContext';
+import { SelectedContextType } from '../../enums/SelectedContextType';
 
 export const useSelectedContext = () => {
   const { selectedContext = SelectedContextType.None } = useParams();
