+++ frontend/dashboard/utils/repoUtils/repoUtils.test.ts
@@ -1,4 +1,4 @@
-import { SelectedContextType } from 'dashboard/context/HeaderContext';
+import { SelectedContextType } from '../../enums/SelectedContextType';
 import { getReposLabel, validateRepoName } from './repoUtils';
 import { textMock } from '@studio/testing/mocks/i18nMock';
 import type { Organization } from 'app-shared/types/Organization';
