+++ frontend/dashboard/utils/userUtils/userUtils.test.ts
@@ -1,4 +1,4 @@
-import { SelectedContextType } from 'dashboard/context/HeaderContext';
+import { SelectedContextType } from '../../enums/SelectedContextType';
 import {
   getOrgNameByUsername,
   getOrgUsernameByUsername,
